refactor(core): extract recompute-icon-hashes from migrate-mantle-state
Move the per-environment icon-hash recomputation pass and its private
helpers into a sibling shell module so migrate-mantle-state stays under
the 300-line cap after the place-support and game-pass slices both
land. The new module owns the IconHashRecomputation result shape, the
ambiguous-warning construction with environment-prefixed mantlePath,
and the readFile -> sha256Hex pipeline; migrateMantleState imports
recomputeIconHashes verbatim. No behavior change.
